@@ -2,14 +2,16 @@ package pl.altkom.springcloud.lab02.feign.projectservice.controller;
 
 import java.util.List;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
@@ -35,6 +37,7 @@ public class ProjectController {
         return projectService.getProject(projectId);
     }
 
+    @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
     public Project createProject(@Valid @RequestBody final CreateProjectRequest request) {
         return projectService.createProject(request);
