@@ -1,6 +1,6 @@
 package pl.altkom.springcloud.lab02.feign.projectservice.controller.model;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 import lombok.Getter;
 
